pyproject.toml
[tool.poetry]
name = "graphql-starter"
version = "0.1.0"
description = "Graphql-starter is a graphql start project with fastapi"
authors = ["huangkai"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = {version = "=0.38.1", extras = ["all"]}
tortoise_orm = "^0.13.8"
python-dotenv = "^0.10.3"
aiomysql = "^0.0.20"
passlib = "^1.7"
[tool.poetry.dev-dependencies]
pytest = "^3.0"

[tool.black]
skip-string-normalization = true

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
verify_ssl = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
